+++ src/Graph.h
@@ -11,7 +11,4 @@
 //This function builds a graph
 void genGraph(ColoredCDBG<UnitigInfo> &cdbg, CCDBG_Build_opt &cdbgOpt);
 
-//This function assigns color ids from the graph to color names in the color set
-void mapColorIds(list<pair<string, size_t>>& colorSet, const ColoredCDBG<UnitigInfo>& g);
-
 #endif