+++ src/IO.cpp
@@ -1,6 +1,7 @@
 #include <getopt.h>
 
 #include "IO.h"
+#include "Sequence.h"
 
 //This function parses the program parameters. Returns false if given arguments are not valid.
 const bool parseArgs(int& nb_args, char** argList, int16_t& prepros, string& filePref, int32_t& s, int32_t& k, int32_t& g,  CCDBG_Build_opt &gOpt, int32_t& t, string& qFile, string& c, uint32_t& m, SrchStrd& strd, bool& r, uint16_t &mscore, int16_t &mmscore, int16_t& X, int32_t &gOpen, int32_t &gExt, uint16_t &nRes, double &lambda, double &lambdaG, double &C, double &Cgap, double &eValLim, bool &isSim, bool &advIdx){
@@ -277,10 +278,13 @@ const bool parseArgs(int& nb_args, char** argList, int16_t& prepros, string& fil
 	return true;
 }
 
-//This function reads in a file in which colors are stored the search will be based on
-const list<pair<string, size_t>> loadSearchColors(const char* filename, uint32_t& nbCols){
+//This function reads in a search color set and maps its colors to the colors of the graph. It terminates the program with error state if either
+//	1. the search color set file cannot be loaded,
+//	2. the search color set file is empty or
+//	3. a color specified in the search color set does not exist in the graph
+const list<pair<string, size_t>> loadSearchColors(const char* filename, const vector<string> &gClrs){
 	string line;
-	pair<string, size_t> color("", -1);//Color id is set to -1 here to avoid false results if color from color search set is not present in the graph<-//TODO: Avoid this by mapping the colors directly within this function!
+	vector<string>::const_iterator i;
 	list<pair<string, size_t>> colorlist;
 
 	//Open the file
@@ -292,16 +296,25 @@ const list<pair<string, size_t>> loadSearchColors(const char* filename, uint32_t
 		exit(EXIT_FAILURE);
 	}
 
-	//Reset counter
-	nbCols = 0;
 	//Read in lines of color set file
 	while(getline(fStr, line)){
-		//Store color name
-		color.first = line;
-		//Store each color
-		colorlist.push_back(color);
-		//Increment counter
-		++nbCols;
+		//Search for color name among all graph color names
+		i = find(gClrs.begin(), gClrs.end(), line);
+
+		//Check if color exists in the graph
+		if(i != gClrs.end()){
+			colorlist.push_back(pair<string, size_t>(line, i - gClrs.begin()));
+		} else{
+			//A color from the search color set does not occur in the graph
+			cerr << "ERROR: Color " << line << " from specified search color set does not occur in the given graph" << endl;
+			exit(EXIT_FAILURE);
+		}
+	}
+
+	//Make sure we could load any colors
+	if(colorlist.empty()){
+		cerr << "ERROR: No colors could be read from search color set file" << endl;
+		exit(EXIT_FAILURE);
 	}
 
 	//Close the file